backend/requirements.txt
# plaid
plaid_python==9.1.0

# server
fastapi==0.85.0
python-dotenv==0.15.0

# other
flake8==3.9.2
pytest==6.2.5
pytest-cov==2.12.1
